+++ src/main/kotlin/core/ext/MessageChannelExtensions.kt
@@ -4,4 +4,4 @@ import core.database.whitelistedSpamChannelIds
 import dev.kord.core.entity.channel.MessageChannel
 
 val MessageChannel.isWhitelistedSpamChannel: Boolean
-    get() = whitelistedSpamChannelIds.contains(id.value)
+    get() = whitelistedSpamChannelIds.contains(id.value.toLong())
